Extract python virtualenv setup and move it to PythonCreator
